Adding SystemD integration LSB override
In certain conditions, such as configuring a dedicated user for
Shellinabox, system tend to crash or not to be reboot proof.

This is because Shellinaboxd starts before essential elements, such as
slapd, and can't recover from failure.

Adding such feature enables Respawn directove to be handled by SystemD.
Boot order is placed after YunoHost's Slapd